[#20240711] Move API endpoints to v1 module
No functional changes.
